feature(security): implementation of security details for phase 1
Modified security configuration to have roles considered for API user access.  Made modifications to code to get the first set of phase1 tdd tests to run.  Also had to make some changes to the tests since I am using the Oauth package to how to mock out the users and roles for the tests.

Git: NA